Pass the Document to the custom AttributConverter when converting attribute is not primitive type.
Currently the enclosed value is passed to the converter, what is good for the primitive types, but in a case  of the entity a list of properties is passed (List<Document>) which is counterintuitive. The Value would also be an option, but the Document is more similar to the TypeReferenceReader.