feat: Add support for converting arrays in ParameterConverter
The code changes in this commit add support for converting arrays in the `ParameterConverter` class. This allows for the conversion of arrays of entities or embedded groups. The `ARRAY` enum value is introduced to handle array conversions, and the `arrayConverter` method is added to determine if the array element type is an entity or embedded group.

Signed-off-by: Maximillian Arruda <dearrudam@gmail.com>